Remove the python-httplib2 dependency
Rather than check the remote VERSION with the 'uri' Ansible module
I decided to just always run rvm update unless you specifically disable
that in the default options.

The benefits of this change is that the role no longer depends on any
packages and CentOS 7 does not have python-httplib2 anyways.

The cons of this change is that the update takes an additional 2-3
seconds to run when it was skipped previously unless needed. However, with the
elimination of quite a few tasks and logic the overall run speed of the
role is faster than before.