Show docker-compose output in Travis to address build timeouts
It takes long enough to build the 6 different distro containers used for
tests that Travis is killing the build during docker-compose up. In the
Travis output, there's a message "No output has been received in the
last 10m0s", which suggests it's killing the job early in part because
we're sending all docker-compose output to /dev/null. Try showing all of
that output in the hopes of improving build reliability.